requirement.txt
streamlit
pandas
numpy
opencv-python-headless
pillow
transformers
torch
ultralytics
folium
streamlit-folium